+++ packages/terra-paginator/tests/wdio/paginator-spec.js
@@ -1,34 +1,31 @@
-const viewports = Terra.viewports('tiny', 'medium', 'large');
-
-describe('Paginator', () => {
+Terra.describeViewports('Paginator', ['tiny', 'medium', 'large'], () => {
   describe('Paginator', () => {
     before(() => browser.url('/#/raw/tests/terra-paginator/paginator/paginator'));
 
-    Terra.it.isAccessible({ viewports });
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.validatesElement();
   });
 
   describe('PaginatorNoPages', () => {
     before(() => browser.url('/#/raw/tests/terra-paginator/paginator/paginator-no-pages'));
 
-    Terra.it.isAccessible({ viewports });
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.validatesElement();
   });
 
   describe('ProgressivePaginator', () => {
     before(() => browser.url('/#/raw/tests/terra-paginator/paginator/progressive-paginator'));
 
-    Terra.it.isAccessible({ viewports });
-    Terra.it.matchesScreenshot({ viewports });
+    Terra.it.validatesElement();
   });
+});
 
+Terra.describeViewports('Controlled Paginator', ['large'], () => {
   describe('Controlled Paginator', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-paginator/paginator/controlled-paginator');
       browser.moveToObject('#root', 0, 900); // move cursor to remove focus styles
     });
 
-    Terra.it.matchesScreenshot('0');
+    Terra.it.validatesElement('0');
 
     it('should toggle page change when the props are updated', () => {
       browser.click('#button-9');
@@ -52,7 +49,7 @@ describe('Paginator', () => {
   describe('Controlled Progressive Paginator', () => {
     before(() => browser.url('/#/raw/tests/terra-paginator/paginator/controlled-progressive-paginator'));
 
-    Terra.it.matchesScreenshot('0');
+    Terra.it.validatesElement('0');
 
     it('should toggle page change when the props are updated', () => {
       browser.click('#button-9');