+++ packages/terra-image/CHANGELOG.md
@@ -9,6 +9,9 @@ Unreleased
 * Added prop imageFit.
 * Added documentation on required peerDependencies
 
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+
 3.11.0  - (July 11, 2019)
 ------------------
 ### Changed