+++ packages/terra-list/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * Added documentation on required peerDependencies
 
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports and use `Terra.it.validatesElement`
+
 4.10.0  - (July 11, 2019)
 ------------------
 ### Changed